langchain-agents/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0
